Add fhir/r5 directory and move helpers
